backend/requirements.txt
# COE892 Decentralized Cryptocurrency - Backend
# Python 3.10+

# REST API (Lab 4)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0

# gRPC (Lab 2)
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# RabbitMQ (Lab 3)
pika>=1.3.0

# CORS
python-multipart>=0.0.6

# Cryptography (transaction signing / verification - proposal)
cryptography>=42.0.0

# Tests (Integration & Testing role)
pytest>=7.0.0
httpx>=0.25.0
